requirements.txt
mysql-connector-python==8.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0